use clickhouse client --multiquery for the rebuild derived tables step
